+++ src/hooks/useSnackbar.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, ReactNode, SetStateAction, createContext, useContext, useState } from 'react';
 
-import { BaseProps } from '../components/common/AlertingSnackbar';
+import { BaseProps } from '../components/common/AlertSnackbar';
 
 type SnackbarContextType = [BaseProps | null, Dispatch<SetStateAction<BaseProps | null>>];
 
@@ -8,9 +8,11 @@ type SnackbarContextType = [BaseProps | null, Dispatch<SetStateAction<BaseProps 
 const SnackbarContext = createContext<SnackbarContextType>([null, () => {}]);
 
 export const SnackbarContextProvider = ({ children }: { children: ReactNode }) => {
-  const [alert, setAlert] = useState<BaseProps | null>(null);
+  const [snackbar, setSnackbar] = useState<BaseProps | null>(null);
 
-  return <SnackbarContext.Provider value={[alert, setAlert]}>{children}</SnackbarContext.Provider>;
+  return (
+    <SnackbarContext.Provider value={[snackbar, setSnackbar]}>{children}</SnackbarContext.Provider>
+  );
 };
 
 export const useSnackbar = () => {
